refactor(verticaltabs): break up components into separate files
i was having a hard time finding my way around. i think this helps
