+++ packages/verticaltabs/src/react/__specs__/index.spec.js
@@ -1,13 +1,11 @@
 import { render } from '@testing-library/react'
 import React from 'react'
 
-import Verticaltab from '../index.js'
+import VerticalTab from '../index.js'
 
 describe('Verticaltab', () => {
   it('renders', () => {
-    const { getByTestId } = render(
-      <Verticaltab data-testid="undertest" />
-    )
+    const { getByTestId } = render(<VerticalTab data-testid="undertest" />)
 
     expect(getByTestId('undertest')).toBeInTheDocument()
   })
@@ -15,9 +13,7 @@ describe('Verticaltab', () => {
   it('forwards refs', () => {
     const ref = React.createRef()
 
-    const { getByTestId } = render(
-      <Verticaltab ref={ref} />
-    )
+    render(<VerticalTab ref={ref} />)
 
     expect(ref.current).not.toBeNull()
   })
